Various UI fixes.
Disabling gradient opacity check-box (ModuleVol) when 2D is active.

Setting active transfer mode when selecting m_activeModule.

Setting a default position (center) in inserted box items.

+++ tomviz/Module.cxx
@@ -349,7 +349,7 @@ void Module::setTransferMode(const int mode)
 int Module::getTransferMode() const
 {
   ///TODO handle detached mode
-  return static_cast<int>(d->m_transferMode);
+  return d->m_transferMode;
 }
 
 } // end of namespace tomviz
